+++ app/src/main/java/com/nougust3/weather/Models/Rain.java
@@ -8,13 +8,13 @@ public class Rain {
 
     @SerializedName("3h")
     @Expose
-    private Integer _3h;
+    private Double _3h;
 
-    public Integer get3h() {
+    public Double get3h() {
         return _3h;
     }
 
-    public void set3h(Integer _3h) {
+    public void set3h(Double _3h) {
         this._3h = _3h;
     }
